+++ src/main/java/com/example/STAT_FetchProjectData/client/FtTokenClient.java
@@ -1,0 +1,7 @@
+package com.example.STAT_FetchProjectData.client;
+
+import com.example.STAT_FetchProjectData.client.dto.Oauth2TokenResponseDto;
+
+public interface FtTokenClient {
+    Oauth2TokenResponseDto getFtToken(String code);
+}
