+++ src/main/java/com/example/STAT_FetchProjectData/config/ClientConfig.java
@@ -2,8 +2,10 @@ package com.example.STAT_FetchProjectData.config;
 
 import com.example.STAT_FetchProjectData.client.FtClient;
 import com.example.STAT_FetchProjectData.client.FtClientImpl;
-import com.example.STAT_FetchProjectData.client.FtClientToken;
-import com.example.STAT_FetchProjectData.client.FtClientTokenImpl;
+import com.example.STAT_FetchProjectData.client.FtTokenClient;
+import com.example.STAT_FetchProjectData.client.FtTokenClientImpl;
+import com.example.STAT_FetchProjectData.repository.FtTokenRepository;
+import com.example.STAT_FetchProjectData.repository.FtMemoryTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -17,5 +19,5 @@ public class ClientConfig {
     public FtClient ftClient() { return new FtClientImpl(tokenConfig); }
 
     @Bean
-    public FtClientToken ftClientToken() { return new FtClientTokenImpl(tokenConfig); }
+    public FtTokenClient ftTokenClient() { return new FtTokenClientImpl(tokenConfig); }
 }
