+++ src/main/java/com/example/STAT_FetchProjectData/service/LoginService.java
@@ -1,17 +1,18 @@
 package com.example.STAT_FetchProjectData.service;
 
-import com.example.STAT_FetchProjectData.client.FtClientToken;
-import com.example.STAT_FetchProjectData.client.FtClientTokenImpl;
+import com.example.STAT_FetchProjectData.client.FtTokenClient;
+import com.example.STAT_FetchProjectData.repository.FtTokenRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
 public class LoginService {
-    private final FtClientTokenImpl ftClientToken;
+    private final FtTokenClient ftTokenClient;
+    private final FtTokenRepository ftTokenRepository;
 
     public String registerAccessToken(String code) {
-        ftClientToken.saveAccessToken(code);
+        ftTokenRepository.saveAccessToken(ftTokenClient.getFtToken(code));
         return "OK";
     }
 }
